requirements.txt
pixivpy-async
python-telegram-bot >= 12.8
bs4